+++ flask_dance/contrib/azure.py
@@ -11,6 +11,7 @@ def make_azure_blueprint(
     client_secret=None,
     *,
     scope=None,
+    offline=False,
     redirect_url=None,
     redirect_to=None,
     login_url=None,
@@ -34,6 +35,8 @@ def make_azure_blueprint(
         client_id (str): The client ID for your application on Azure AD.
         client_secret (str): The client secret for your application on Azure AD
         scope (str, optional): comma-separated list of scopes for the OAuth token
+        offline (bool): Whether to request offline access for the OAuth token.
+            Defaults to False, but must be True to receive a refresh token.
         redirect_url (str): the URL to redirect to after the authentication
             dance is complete
         redirect_to (str): if ``redirect_url`` is not defined, the name of the
@@ -76,6 +79,15 @@ def make_azure_blueprint(
     :returns: A :doc:`blueprint <flask:blueprints>` to attach to your Flask app.
     """
     scope = scope or ["openid", "email", "profile", "User.Read"]
+    token_url = f"https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token"
+    authorization_url = (
+        f"https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize"
+    )
+    auto_refresh_url = None  # see #161
+    if offline:
+        scope = [scope] if not isinstance(scope, list) else scope
+        scope.append("offline_access")
+        auto_refresh_url = token_url
     authorization_url_params = {}
     if login_hint:
         authorization_url_params["login_hint"] = login_hint
@@ -90,8 +102,9 @@ def make_azure_blueprint(
         client_secret=client_secret,
         scope=scope,
         base_url="https://graph.microsoft.com",
-        authorization_url=f"https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize",
-        token_url=f"https://login.microsoftonline.com/{tenant}/oauth2/v2.0/token",
+        authorization_url=authorization_url,
+        token_url=token_url,
+        auto_refresh_url=auto_refresh_url,
         redirect_url=redirect_url,
         redirect_to=redirect_to,
         login_url=login_url,