+++ tests/contrib/test_azure.py
@@ -39,6 +39,31 @@ def test_blueprint_factory():
         azure_bp.token_url
         == "https://login.microsoftonline.com/common/oauth2/v2.0/token"
     )
+    assert (
+            azure_bp.auto_refresh_url is None
+    )
+
+
+def test_blueprint_factory_offline():
+    azure_bp = make_azure_blueprint(
+        client_id="foo", client_secret="bar", scope="user.read", redirect_to="index", offline=True
+    )
+    assert isinstance(azure_bp, OAuth2ConsumerBlueprint)
+    assert azure_bp.session.scope == ["user.read", "offline_access"]
+    assert azure_bp.session.base_url == "https://graph.microsoft.com"
+    assert azure_bp.session.client_id == "foo"
+    assert azure_bp.client_secret == "bar"
+    assert (
+            azure_bp.authorization_url
+            == "https://login.microsoftonline.com/common/oauth2/v2.0/authorize"
+    )
+    assert (
+            azure_bp.token_url
+            == "https://login.microsoftonline.com/common/oauth2/v2.0/token"
+    )
+    assert (
+            azure_bp.auto_refresh_url is azure_bp.token_url
+    )
 
 
 def test_blueprint_factory_with_domain_hint():